requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
discord.py==1.7.1
idna==3.1
multidict==5.1.0
pycparser==2.20
PyYAML==5.4.1
typing-extensions==3.7.4.3
yarl==1.6.3
